LPD-99112 refactor: extract the reusable space select field
